Keep original (0-255) value longer

+++ src/main/java/com/getpcpanel/device/PCPanelRGBUI.java
@@ -82,8 +82,7 @@ public class PCPanelRGBUI extends Device {
 
     @Override
     public int getKnobRotation(int knob) {
-        //noinspection NumericCastThatLosesPrecision
-        return Util.rotationToAnalogValue((int) ((Region) knobs[knob].getGraphic()).getChildrenUnmodifiable().get(3).getRotate());
+        return Util.rotationToAnalogValue(((Region) knobs[knob].getGraphic()).getChildrenUnmodifiable().get(3).getRotate());
     }
 
     private void initLabel() {
